Add CreatedSince and UpdatedSince components
