+++ src/components/CreatedSince/index.ts
@@ -1,0 +1,1 @@
+export { CreatedSinceContainer as CreatedSince } from "./container";
