+++ src/components/TimeHumanizer/index.tsx
@@ -1,1 +1,23 @@
-export { TimeHumanizerContainer as TimeHumanizer } from "./container";
+import React, { FunctionComponent } from "react";
+import classNames from "classnames";
+import moment from "moment";
+import "moment/locale/es";
+import styles from "./styles.module.scss";
+
+export const TimeHumanizer: FunctionComponent<ITimeHumanizerProps> = (
+  {
+    className,
+    since,
+    labelPrefix
+  }
+) => (
+  <p className={classNames(styles.time, className)}>
+    {`${labelPrefix} ${moment(since).fromNow()}`}
+  </p>
+);
+
+interface ITimeHumanizerProps {
+  className?: string;
+  since: string;
+  labelPrefix: string;
+}
