+++ src/components/UpdatedSince/index.ts
@@ -1,0 +1,1 @@
+export { UpdatedSinceContainer as UpdatedSince } from "./container";
