+++ src/pages/Admin/Home/Dashboard/TaskDetail/MainTitle/index.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from "react";
-import { TimeHumanizer } from "$components/TimeHumanizer";
+import { CreatedSince } from "$components/CreatedSince";
 import styles from "./styles.module.scss";
 
 export const MainTitle: FunctionComponent<IMainTitleProps> = (
@@ -13,7 +13,7 @@ export const MainTitle: FunctionComponent<IMainTitleProps> = (
       <p className={styles.title}>
         {title}
       </p>
-      <TimeHumanizer since={createdAt} type="create"/>
+      <CreatedSince date={createdAt} />
     </div>
   </div>
 );
